Create chrome-style editor tabs
Tabs take the role of the main app router - they are responsible for
mounting the correct component, any sagas and the delegating to the
correct reducer when using the relayActionToTab meta-action. The tabs
perform this functionality using a combination of a higher-order
reducers to control reducer delegation, Redux middleware to control saga
activity and a higher-order component to connect a React component to
the store and its tab state (by wrapping the default react-redux
connect).

Sagas gain a number of new effects to handle tab-related actions. These
are simply wrappers over the original effects, except they are able to
filter actions based on the tab ID, allowing a tab to listen only for
the events that concern it (effectively multiplexing actions).